app/requirements.txt
flask==2.3.2
gunicorn==21.2.0
PyYAML==6.0.2
